+++ src/App.js
@@ -1,13 +1,13 @@
 import 'normalize.css';
-import { useState, useEffect } from 'react';
+import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useSelector, useDispatch } from 'react-redux';
-import toast, { Toaster } from 'react-hot-toast';
-import { requestForToken, onMessageListener } from './firebase';
 
 import useAxios from './hooks/useAxios';
 import Authorized from './routes/Authorized';
 import Unauthorized from './routes/Unauthorized';
+
+import Notification from './components/Notification';
 
 import { saveUserInfo } from './store/userSlice';
 
@@ -16,18 +16,6 @@ function App() {
   const navigate = useNavigate();
   const userInformation = useSelector((state) => state.user.userInformation);
   const isLoggedIn = localStorage.getItem('isLoggedIn') === 'true';
-  const [notification, setNotification] = useState({ title: '', body: '' });
-  const notify = () => toast(<ToastDisplay />);
-  function ToastDisplay() {
-    return (
-      <div>
-        <p>
-          <b>{notification?.title}</b>
-        </p>
-        <p>{notification?.body}</p>
-      </div>
-    );
-  }
 
   useEffect(() => {
     const setUserInformation = async () => {
@@ -46,16 +34,11 @@ function App() {
     }
   }, [userInformation]);
 
-  useEffect(() => {
-    // const Notification = () => {
-    requestForToken();
-    // };
-  }, []);
-
   return (
     <>
       {isLoggedIn && userInformation && <Authorized />}
       {!isLoggedIn && <Unauthorized />}
+      <Notification />
     </>
   );
 }
